+++ packages/vite-plugin-cloudflare/src/index.ts
@@ -13,11 +13,7 @@ import {
 	matchAdditionalModule,
 } from "./additional-modules";
 import { hasAssetsConfigChanged } from "./asset-config";
-import {
-	checkPublicDir,
-	getImportedAssetPaths,
-	loadViteManifest,
-} from "./build";
+import { createBuildApp } from "./build";
 import {
 	createCloudflareEnvironmentOptions,
 	initRunners,
@@ -167,157 +163,7 @@ export function cloudflare(pluginConfig: PluginConfig = {}): vite.Plugin[] {
 					builder: {
 						buildApp:
 							userConfig.builder?.buildApp ??
-							(async (builder) => {
-								const clientEnvironment = builder.environments.client;
-								assert(clientEnvironment);
-								const clientBuildDirectory = path.resolve(
-									builder.config.root,
-									clientEnvironment.config.build.outDir
-								);
-								const defaultHtmlPath = path.resolve(
-									builder.config.root,
-									"index.html"
-								);
-								const hasClientEntry =
-									clientEnvironment.config.build.rollupOptions.input ||
-									fs.existsSync(defaultHtmlPath);
-
-								if (resolvedPluginConfig.type === "assets-only") {
-									if (hasClientEntry) {
-										await builder.build(clientEnvironment);
-									} else {
-										const hasPublicAssets = checkPublicDir(builder.config);
-
-										if (hasPublicAssets) {
-											const tempEntryName = "__cloudflare_temp_entry__";
-											clientEnvironment.config.build.rollupOptions = {
-												input: "virtual:__cloudflare_temp_entry__",
-												logLevel: "silent",
-												output: {
-													entryFileNames: tempEntryName,
-												},
-											};
-
-											await builder.build(clientEnvironment);
-
-											const tempEntryPath = path.join(
-												clientBuildDirectory,
-												tempEntryName
-											);
-											fs.unlinkSync(tempEntryPath);
-										}
-									}
-
-									return;
-								}
-
-								const workerEnvironments = Object.keys(
-									resolvedPluginConfig.workers
-								).map((environmentName) => {
-									const environment = builder.environments[environmentName];
-
-									assert(
-										environment,
-										`${environmentName} environment not found`
-									);
-
-									return environment;
-								});
-
-								await Promise.all(
-									workerEnvironments.map((environment) =>
-										builder.build(environment)
-									)
-								);
-
-								const { entryWorkerEnvironmentName } = resolvedPluginConfig;
-								const entryWorkerEnvironment =
-									builder.environments[entryWorkerEnvironmentName];
-								assert(entryWorkerEnvironment);
-								const entryWorkerBuildDirectory = path.resolve(
-									builder.config.root,
-									entryWorkerEnvironment.config.build.outDir
-								);
-								const entryWorkerManifest = loadViteManifest(
-									entryWorkerBuildDirectory
-								);
-								const importedAssetPaths =
-									getImportedAssetPaths(entryWorkerManifest);
-
-								if (hasClientEntry) {
-									await builder.build(clientEnvironment);
-								} else {
-									const hasAssets =
-										importedAssetPaths.size || checkPublicDir(builder.config);
-
-									if (hasAssets) {
-										const tempEntryName = "__cloudflare_temp_entry__";
-										clientEnvironment.config.build.rollupOptions = {
-											input: "virtual:__cloudflare_temp_entry__",
-											logLevel: "silent",
-											preserveEntrySignatures: "strict",
-											output: {
-												entryFileNames: tempEntryName,
-											},
-										};
-
-										await builder.build(clientEnvironment);
-
-										const tempEntryPath = path.resolve(
-											builder.config.root,
-											clientEnvironment.config.build.outDir,
-											tempEntryName
-										);
-										fs.unlinkSync(tempEntryPath);
-									} else {
-										const entryWorkerConfigPath = path.join(
-											entryWorkerBuildDirectory,
-											"wrangler.json"
-										);
-										const workerConfig = JSON.parse(
-											fs.readFileSync(entryWorkerConfigPath, "utf-8")
-										);
-										workerConfig.assets = undefined;
-										fs.writeFileSync(
-											entryWorkerConfigPath,
-											JSON.stringify(workerConfig)
-										);
-									}
-								}
-
-								const movedAssetPaths: string[] = [];
-
-								for (const assetPath of importedAssetPaths) {
-									const src = path.join(entryWorkerBuildDirectory, assetPath);
-									const dest = path.join(clientBuildDirectory, assetPath);
-
-									if (fs.existsSync(dest)) {
-										fs.unlinkSync(src);
-									} else {
-										const destDir = path.dirname(dest);
-
-										if (!fs.existsSync(destDir)) {
-											fs.mkdirSync(destDir, { recursive: true });
-										}
-
-										fs.renameSync(src, dest);
-										movedAssetPaths.push(dest);
-									}
-								}
-
-								if (movedAssetPaths.length) {
-									builder.config.logger.info(
-										[
-											`${colors.green("✓")} ${movedAssetPaths.length} asset${movedAssetPaths.length > 1 ? "s" : ""} moved from ${entryWorkerEnvironmentName} to client build output.`,
-											...movedAssetPaths.map((assetPath) =>
-												colors.dim(
-													path.relative(builder.config.root, assetPath)
-												)
-											),
-										].join("\n")
-									);
-								}
-							}),
+							createBuildApp(resolvedPluginConfig),
 					},
 				};
 			},
@@ -584,19 +430,19 @@ export function cloudflare(pluginConfig: PluginConfig = {}): vite.Plugin[] {
 				});
 			},
 		},
-		// Plugin to provide a dummy client entry file
+		// Plugin to provide a fallback entry file
 		{
-			name: "vite-plugin-cloudflare:temp-client-entry",
-			applyToEnvironment(environment) {
-				return environment.name === "client";
-			},
+			name: "vite-plugin-cloudflare:fallback-client-entry",
+			// applyToEnvironment(environment) {
+			// 	return environment.name === "client";
+			// },
 			resolveId(source) {
-				if (source === "virtual:__cloudflare_temp_entry__") {
-					return `\0virtual:__cloudflare_temp_entry__`;
+				if (source === "virtual:__cloudflare_fallback_entry__") {
+					return `\0virtual:__cloudflare_fallback_entry__`;
 				}
 			},
 			load(id) {
-				if (id === "\0virtual:__cloudflare_temp_entry__") {
+				if (id === "\0virtual:__cloudflare_fallback_entry__") {
 					return ``;
 				}
 			},
